metrics: Fix custom initialization
New... functions in pkg/metrics take an initialization function as an argument,
in case custom initialization is needed. However, such function passed as part
of a metric definition can't reference this metric - this would mean an
initialization cycle. This commit changes the initialization functions
signatures to take an inner prometheus metric as an argument, so that we can
actually initialize it and avoid a cycle.

Signed-off-by: Anna Kapuscinska <anna@isovalent.com>