tetra: switch CliRun to use NewClient instead of DialContext
DialContext got deprecated, so we switch to NewClient that does not
connect to the server, it waits for the first RPC to be called to
establish a connection, thus the retry mechanism at connect no longer
make sense. It could be reimplemented as a retry helper for the RPC
call.

Signed-off-by: Mahe Tardy <mahe.tardy@gmail.com>